Migrate remaining actions as commands
Mostly found audio actions. Unfortunately, it was necessary to remove
most of its funcionality, as it was using background goroutines that had
concurrent access to the application state. And there was no control
over concurrent changes in functions like fade in or fade out.

For now, as we just need the ability to play music tracks, only the play
and stop functions have been maintained as commands.