Introduce costume resources
Having animations directly associated to actors is not convenient. We
might need to swap the appearance of an actor, and also it is simpler
to have the animations encapsulated under the same resource type.

This change introduces a Costume resource following the same concept as
in SCUMM. The Costume is a spritesheet with some animations defined on
it for standing, walking, speaking, and some other custom ones.

To fit this change in the resource model, the Actor and the SpriteSheet
are no longer considered resources. Actors are just identities of people
who act (walk, speak, stand) and have their own inventory. Actors can
use a costume to show visually their actions. If no costume is selected,
the actor is invisible (as in SCUMM).