Fix wrong way of loading Lua code
The lua.DoString function is a crap. It sets the string passed as
argument as source for debugging info. So you cannot tell where the
error is when something goes wrong.

This change uses the State.Load function instead. Which admits a chunk
name to be passed as argument. Now any error is reported indicating the
script resource where it is detected and the line.