Ensure costume speak after movement
So far, if the actor moves while it was speaking, the dialog is still
shown but the actor no longer speaks (the costume is idle).

This change makes `Actor` aware of the current `Dialog` and it uses the
speak costume animation while standing if there is an ongoing dialog.

Unfortunately, we cannot do this for walking because we cannot combine
this animation with speaking.